[backend] Refactor to extract attachment props from engine.ts
